Presentation corrected result (type)
proper presentation, proper json output, proper result for
simulated annealing and other refactoring.

SA made adjustment to allow proper timeout control and iteration
run settings.

preliminary test shows 1000 run yields 1562 energy and 2000 runs
yield 662, an 110% improvement in terms of score. [simulated
annealing]

+++ src/main/java/model/CandidateSolution.java
@@ -74,10 +74,10 @@ public class CandidateSolution implements Comparable<CandidateSolution> {
     }
 
     /**
-     * @return
+     * @return inverse of the energy score
      */
     public double getFitness(){
-        return 1/getEnergy();
+        return 1/(double)getEnergy();
     }
 
     @Override
